ensure delete call findById with correct params

+++ backend/todolist/src/main/java/com/koglin/todolist/infra/database/repositories/TaskRepositoryJpa.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TaskRepositoryJpa extends JpaRepository<TaskEntity, Integer> {
+public interface TaskRepositoryJpa extends JpaRepository<TaskEntity, Long> {
 }
